Fixed build issue for android .apks on windows.
If more than one process calls "android project update" at the same time
an internal error occurs and causes it to crash, and the build
steps to not complete. This change, on windows, forces all
calls to android project update to be serialized. This effectively
turns that part of the build into single-process, but only that piece
of the build the rest can be scaled correctly.

Change-Id: I41713e2fcc48e6bb821ee0f1dbc4f69a74ed6587
Reviewed-on: https://wn-project.dnsalias.net/12
Reviewed-by: devpow112 <devon.f.powell@gmail.com>
Tested-by: devpow112 <devon.f.powell@gmail.com>