Added WNHashSet.
If we decide to optimize in the future, we can, but for now
this just wraps a hash_map with a value_type of
an empty struct. This may not be perfectly ideal, since
the struct will end up taking some amount of space,
but in practice I do not expect this type to get used
a lot, or stay around for a long time.

Change-Id: I64d67946ef67757ffb789e02732cbd1a8ba089fa
Reviewed-on: https://wn-project.dnsalias.net/48
Reviewed-by: devpow112 <devon.f.powell@gmail.com>
Tested-by: devpow112 <devon.f.powell@gmail.com>