Greatly improved android debug support.
Works on multiple android devices now.
By providing --gdb_command to the debugger you can
use a non-standard debugger. For example,
I have successfully used:
--gdb_command="C:\cygwin\bin\cgdb.exe -d $gdb $args".

Notes: On windows gdb will do the wrong thing when encountering
ctrl-C. (Usually just crashes the process). To combat this,
a command-line of the form:

/path/to/adb shell run-as 'package_name' kill -5 pid is printed
to the output.
You can run that command in a terminal to cause the program
to stop execution.

Change-Id: Ica6e2c04ee5a5da99944ff3c760d78345e48521f