Added WNTesting library.
This allows us to only have a single place to handle testing.
This defines WNMain for the testing module, and sets up
gtest properly. On android it adds a custom gtest reporter
to output the data to the android log, which is then picked up
by the android python scripts.

This also removes the param() object from scripting,
it added nothing but complication.

Lastly added support for Android 6.0, which
changed the logging format. Now both pre and post
6.0 devices have correct logging output.

Change-Id: Ifd4b263670c9b4ea24d5e8fd5312335c9c0ce20d
Reviewed-on: https://wn-project.dnsalias.net/gerrit/75
Try: Andrew Woloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>