Added upload heap type to graphics.
This also adds a write-only value, and write-only range to go
along with it.

Even though reading the data is technically possible, the performance
will likely be very poor in many circumstances. The writes will
bypass the caches and write directly to main memory in order to
maintain the coherency requirements. This means that in order
to read we have to read back into L1 cache, after every write.

Because of this guarantee, we do not have to do anything when we flush.
If we expand the abilities of this heap in the future, then we may
have to actually do something when flushing.

Change-Id: I73eb8b921d8ae1c3ccb3675babd50680976d29b6
Reviewed-on: https://wn-project.dnsalias.net/gerrit/218
Try: Andrew Woloszyn <awoloszyn@google.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>
Tested-by: WNBot <wn.bot.1@gmail.com>