Added download-heap to WNGraphics.
Also changed the allocation/free model of heaps.
heap_buffers are returned from a call to get_range,

For upload_heaps the memory is flushed to the GPU
when the heap_buffer is released.

For download_heaps the CPU cache is invalidated,
and the data is acquired from the GPU when the
heap_buffer is created.

Change-Id: I12d593963b257ea47a24403abb061e0112a9a7b4
Reviewed-on: https://wn-project.dnsalias.net/gerrit/224
Try: Andrew Woloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>