Added destructor generation to WNScripting.
They currently are empty, and do not get called. They are
empty because they cannot contain constructible objects so
there is literally nothing for them to do. They are not called
because there is still a bit of work to do to handle temporaries.

Change-Id: Ib764fa9c9c4505b8e10ae50be7581766a603f87a
Reviewed-on: https://wn-project.dnsalias.net/gerrit/246
Try: Andrew Woloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>