Added arbitrary scoping to WNScript.
This allows treats {} as instructions. This can have
future implications like statements such as

for(;; { a  bunch of arbitrary stuff} ) {
}
which should be functionally fine.

This was done primarily to enable a if-rewriting pass
that will needed when we have to handle temporary objects
and if statements.

Change-Id: I8c0104105b2e1f25c6008b8f47412be6378219fb
Reviewed-on: https://wn-project.dnsalias.net/gerrit/247
Try: Andrew Woloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>