Finished normalizing if statements for the C generator.
This finished the process of linearizing if/elseif/else
statements into a set of if statements. The only
case where there is an else statement is where we have
and if/else witout intermediate esleif statements.
This is because it is impossible to have intermediates
there.

Change-Id: I72fbbd0bcaa2ef333c2564de4df5115c0b1ef563
Reviewed-on: https://wn-project.dnsalias.net/gerrit/250
Try: Andrew Woloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>