Allowed arbitrary expressions as instructions.
Now you can write
int foo() { 2 * 3; return 4; }

This is not terribly useful in and of itself, but it
allows expressions with side-effects to be used as instructions.

For example function_calls which are expressions often have
side-effects (destructors for example), as well as expressions
like x++. It would be nice to allow them as arbitrary instructions.

Change-Id: I5873502d0dfeafd775885118c6bc68a9cad76da1
Reviewed-on: https://wn-project.dnsalias.net/gerrit/255
Try: Andrew Woloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: Devon Powell <devon.f.powell@gmail.com>