Added WNSurface to graphics.
This splits surface creation/handling out from swapchain
creation/handling. This is more in line with the Vulkan menatily
of surface/swapchains.

It should make it easier to determine if you want to use a particular
adapter since you don't have to create a full device in order to
test surfaces.

Change-Id: Id1a51011121c080a63adb6dd3dacd82ae26c949e
Reviewed-on: https://wnproject.org/gerrit/22
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: devpow112 <devon.f.powell@gmail.com>