Fix offset of Windows window.
The previous fix, fixed one of the issues with the Window,
specifically the window was being put in totally the
wrong spot. However, now the client-area of the window
is getting set to the provided position. This updates
it so the top-left corner is at the right position, and
the client-area is the right size.

Change-Id: I5ce6e9f7a0fb9abf93a9eb41d4046b6419ed4f98
Reviewed-on: https://wnproject.org/gerrit/53
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: devpow112 <devon.f.powell@gmail.com>