Fixed the flakey routed networking tests.
The problem was that were not waiting sufficiently well. What was
happning was that the semaphore was waiting on the signals, however
the signals were before the routed buffers were destroyed.
In some rare circumstances (1% of the time) the outer
thread was finishing, deleting the job/pool before the
routed buffers were cleared. Clearing the buffers then caused
all sorts of havok.

Added a shutdown() method to the job pool that waits until
all jobs are actually done. If this test was inside the
application framework we could have waited on a signal first,
which would have worked.

Change-Id: Ic07d06615b3b1ba21d0538e194db6914f1451b3c
Reviewed-on: https://wnproject.org/gerrit/54
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: devpow112 <devon.f.powell@gmail.com>