Add first pass of static arrays to scripting.
Tests are still missing for objects in static arrays.
Also shared objects in static arrays.
Also have to handle static arrays in structs/classes.

We have to be a bit careful about putting non-shared
objects in arrays.

For now we will just disable putting classes in
structs. When we start allowing it, at that point
we can define a property on inheritance which defines
whether a type is "compatible".

Example of things that are compatible:
Overloading methods.
Initialing integral types with a different value.

Things that are not compatible.
Adding member variables.
Initializing an array with a different size.
Initializing any child class with a non-compatible
alternative.

Change-Id: I3cae404ce7d97bbd9d613b1d99796fcff1d3657d
Reviewed-on: https://wnproject.org/gerrit/233
Reviewed-by: devpow112 <devon.f.powell@gmail.com>
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>