Added slices to Scripting.
This has a few changes, they are value-types, and passed by
value, but we do not handle passing structs on the stack, so
they are actually passed by reference (and copied onto the stack).

This lets us create 1d slices, and interact with N-d slices (which
can still be created in c++).

Change-Id: I8f3f3f6f3a92fec267d1e67707c188ec3581e762
Reviewed-on: https://wnproject.org/gerrit/255
Reviewed-by: devpow112 <devon.f.powell@gmail.com>
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>