Update LLVM Version.
This remove some of the terrible hacks we needed in the LLVM CMake
files.

We still have one hack, but its only 2 lines now. Because of this
we also update our Android NDK version to r18. Lastly we updated
the NDK STL from gnustl to libc++_static. This was because of some
incompatibilities in llvm that we didnt want to have to work around.

This necessitated a couple of changes in a few places due to compiler
changes and library change, but nothing major.

Change-Id: I8d07725d6ca840b5bb2147b303028504e648c2ca
Reviewed-on: https://wnproject.org/gerrit/263
Try: AWoloszyn <awoloszyn@google.com>
Tested-by: WNBot <wn.bot.1@gmail.com>
Reviewed-by: devpow112 <devon.f.powell@gmail.com>