Add resources to WNScripting.
Resources are declared by the compiler, and can be used with
a @ResourceName("ResourceParameter").

A resource is registered with the scripting engine, and
provides a custom type. A callback is made when hte resource is
processed.

The resource allows for additional custom includes, as well as
custom processing of specific extensions. This means
that a given resource extension can custom modify files as needed.

Change-Id: I75532510eeb00d5013bbb9fe92f73d29b772dc0d
Reviewed-on: https://wnproject.org/gerrit/319
Try: AWoloszyn <awoloszyn@google.com>
Reviewed-by: devpow112 <devon.f.powell@gmail.com>
Tested-by: WNBot <wn.bot.1@gmail.com>