Add input handling to ui system.
This adds basic input, still missing textures, also
still missing temporary geometry. Both of those should be easy.

This "fixes" a bug with the range parition where merging nodes
was not correct. We were dropping parts of the free-list and
overwriting data, it was bad.

The "fix" is to remove the merging behavior, so now the range
partition may easily get fragmented. To be fixed (#168)