Textures in engine (pt1)
This adds a TextureDescription object that holds all of the
conversion information for a texture. This is used
as a source for actually converting textures for loading.

For now it is not actually used, but the framework is there.

Pt2 will consist of actually having the texture manager actually
start loading PNG textures with this layout/format.