Add Tracy Integration
This is disabled by default (and all behavior is unchanged).
But if you -DWN_ENABLE_TRACY=ON you will get Tracy support
turned on. You can build/run the UI from the tracyf directory,
it is not done for you. But once that is done you can
connect to any running program.