Add ability to skip parts of the build (#304)
This analyzes changes and determines if a code format check and/or build is
needed. Helpful when making none-build related changes. Should help reduce the
amount of building we have top do. This will still always run format checks and
a full build on a merge to master. This requires a change to required statuses.
Resolves #253.