Another update to simplify actors.
This simply creates an extra function that primes the
actor call.

So instead of having to inject a bunch of actor related
business at the call-site, we can now just call the
wrapper function that actually handles allocating the
actor data, and it acts like a "normal" function call.

Furthermore this sets us up for 2 things
  1) being able to defer an actor update until a future time
     (# of frames)
  2) Calling actions from C++. This will be important very soon
     as the UI model is firming up.