Another update for actors. (#580)
This update provides a somewhat final update for how actors should
work. There might be some updates in how the renderer interacts with
actors, but for the most part this is the completed actor model.

The only thing missing is being able to defer an actor call for
some amount of time.