Remove librocket, add RMLUi
This replaces librocket with RmlUI which is actively maintained
(as of 2022) and has some nice features.

Instead of modifying the repository directly, this also just
uses it as a submodule with a patch.

The main feature of this patch is to remove the global state
(either method-local statics or proper global statics) which
prevent this from

a) running on multiple threads at once
b) having multiple different instances running concurrently in an
application.

This does however add a few new dependencies (which give us some
nice features). Specifically it uses freetype instead of stbfont
meaning one less patch for us to maintain, it also incorporates
rlottie for animated vector art inside RML elements.